chore: update path names for renamed type file playlistTypes
